refactor(elixir): isolate sanitizer state machine from parser
